feat: reorganize files
* Moving github event fetchers to `./src/events`
* Splitting up `./src/db/` into multiple database operators
* Added a shell oss-directory action
