+++ indexer/package.json
@@ -38,13 +38,13 @@
   "devDependencies": {
     "@types/jest": "^29.5.3",
     "@types/lodash": "^4.14.196",
-    "@types/node": "^20.4.5",
+    "@types/node": "^20.4.8",
     "@types/npm-registry-fetch": "^8.0.4",
     "@types/yargs": "^17.0.24",
     "jest": "^29.6.2",
-    "prettier": "^3.0.0",
-    "prisma": "^5.1.0",
-    "supabase": "^1.82.2",
+    "prettier": "^3.0.1",
+    "prisma": "^5.1.1",
+    "supabase": "^1.82.6",
     "ts-jest": "^29.1.1",
     "ts-jest-resolver": "^2.0.1",
     "ts-node": "^10.9.1",
@@ -52,7 +52,7 @@
     "typescript": "^5.1.6"
   },
   "dependencies": {
-    "@prisma/client": "^5.1.0",
+    "@prisma/client": "^5.1.1",
     "chalk": "^5.3.0",
     "dayjs": "^1.11.9",
     "dotenv": "^16.3.1",
@@ -60,6 +60,7 @@
     "graphql-request": "^6.1.0",
     "lodash": "^4.17.21",
     "npm-registry-fetch": "^15.0.0",
+    "oss-directory": "^0.0.2",
     "ts-adt": "^2.1.2",
     "winston": "^3.10.0",
     "yaml": "^2.3.1",
