+++ indexer/src/actions/github/upsertOrg/createEventPointersForRepo.ts
@@ -5,7 +5,7 @@ import {
   Prisma,
   Artifact,
 } from "@prisma/client";
-import { getRepositoryCreatedAt } from "../../../utils/github/getRepositoryCreatedAt.js";
+import { getRepositoryCreatedAt } from "../../../events/github/getRepositoryCreatedAt.js";
 import { GithubIssueClosedInterface } from "../fetch/issueClosed.js";
 import { GithubIssueFiledInterface } from "../fetch/issueFiled.js";
