+++ indexer/src/events/github/getGithubPointer.ts
@@ -1,8 +1,8 @@
 import { Artifact } from "@prisma/client";
 import { GithubFetchArgs } from "../../actions/github/fetch/issueFiled.js";
 import { GithubEventPointer } from "../../actions/github/upsertOrg/createEventPointersForRepo.js";
-import { EventType, prisma } from "../../db/prisma-client.js";
-import { InvalidInputError } from "../error.js";
+import { EventType, prisma } from "../../db/events.js";
+import { InvalidInputError } from "../../utils/error.js";
 
 export const getArtifactName = (org: string, repo: string) => `${org}/${repo}`;
