+++ package.json
@@ -33,7 +33,7 @@
     "prettier": "^3.0.0",
     "turbo": "^1.10.12"
   },
-  "packageManager": "pnpm@8",
+  "packageManager": "pnpm@8.6.12",
   "engines": {
     "node": "18.x",
     "pnpm": "8"
